Move WiFi commands to separate module and fix circular imports
- Create wifi_commands.py for all WiFi-related commands
- Remove WiFi commands from system_commands.py
- Update commands.py to import wifi_commands
- Update wifi.py to import system_commands directly
- Use shell = system_commands.shell for cleaner code
