+++ gonotego/command_center/commands.py
@@ -11,6 +11,7 @@ from gonotego.command_center import note_commands  # noqa: F401
 from gonotego.command_center import settings_commands  # noqa: F401
 from gonotego.command_center import system_commands  # noqa: F401
 from gonotego.command_center import twitter_commands  # noqa: F401
+from gonotego.command_center import wifi_commands  # noqa: F401
 from gonotego.command_center import registry
 
 register_command = registry.register_command
